requirements.txt
streamlit
pandas=2.1.1
numpy==2.0.2
tensorflow
scikit-learn
pillow
openpyxl
joblib==1.5.2
